Moving Page helpers into Page dir
