+++ site/src/components/Page/index.tsx
@@ -1,1 +1,3 @@
 export * from "./Footer"
+export * from "./RequireAuth"
+export * from "./AuthAndNav"
